Migrate documentation URL to shared.
